+++ platform/api/aws/images.go
@@ -15,6 +15,7 @@
 package aws
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/url"
 	"strings"
@@ -331,6 +332,28 @@ func (a *API) CreateHVMImage(snapshotID string, diskSizeGiB uint, name string, d
 	params.EnaSupport = aws.Bool(true)
 	params.SriovNetSupport = aws.String("simple")
 	return a.createImage(params)
+}
+
+func (a *API) RemoveLaunchPermission(imageid string) ([]byte, error) {
+	mod := &ec2.ModifyImageAttributeInput{
+		ImageId: &imageid,
+		LaunchPermission: &ec2.LaunchPermissionModifications{
+			Remove: []*ec2.LaunchPermission{
+				{
+					Group: aws.String("all"),
+				},
+			},
+		},
+	}
+	resp, err := a.ec2.ModifyImageAttribute(mod)
+	if err != nil {
+		return []byte{}, fmt.Errorf("modifying image attribute: %w", err)
+	}
+	out, err := json.Marshal(resp)
+	if err != nil {
+		return []byte{}, fmt.Errorf("marshaling the API response: %w", err)
+	}
+	return out, nil
 }
 
 func (a *API) deregisterImageIfExists(name string) error {
@@ -674,6 +697,24 @@ func (a *API) describeImage(imageID string) (*ec2.Image, error) {
 		return nil, fmt.Errorf("couldn't describe image %q: %v", imageID, err)
 	}
 	return describeRes.Images[0], nil
+}
+
+func (a *API) GetImageLastLaunchedTime(imageID string) (time.Time, error) {
+	resp, err := a.ec2.DescribeImageAttribute(&ec2.DescribeImageAttributeInput{
+		Attribute: aws.String(ec2.ImageAttributeNameLastLaunchedTime),
+		ImageId:   aws.String(imageID),
+	})
+	if err != nil {
+		return time.Time{}, err
+	}
+	if resp.LastLaunchedTime == nil || resp.LastLaunchedTime.Value == nil {
+		return time.Time{}, nil
+	}
+	lastLaunchedTime, err := time.Parse(time.RFC3339Nano, *resp.LastLaunchedTime.Value)
+	if err != nil {
+		return time.Time{}, err
+	}
+	return lastLaunchedTime, nil
 }
 
 // GetImagesByTag returns all EC2 images with that tag